requirements.txt
torch
rembg[gpu]
icecream
Pillow
